+++ plugins/gh-datainmap/CHANGELOG.md
@@ -1,4 +1,8 @@
+
+- Check voor weergave pointer cursor aangepast.
+
 **1.4.0 (2020-03-26)**
+
 - Mogelijkheid om bij instellingen de startlocatie te selecteren via een kaart.
 
 **1.3.5 (2020-03-18)**